Refactor tests for readability
